pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sleepy_factory"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "rich>=13.7",
]

[project.scripts]
sf = "sleepy_factory.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sleepy_factory*"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "pytest>=9.0.2",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
